LLM_project/training/pyproject.toml
[project]
name = "training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "accelerate>=1.13.0",
    "datasets>=4.5.0",
    "peft>=0.17.1",
    "pillow>=12.1.1",
    "pyyaml>=6.0.3",
    "transformers",
    "trl>=0.24.0",
    "wandb>=0.25.1",
]

[tool.uv.sources]
transformers = { git = "https://github.com/huggingface/transformers.git" }